minor checkpoint #4
