+++ client/src/routes/router.jsx
@@ -6,6 +6,9 @@ import { AnimatePresence } from "framer-motion";
 // HOMR
 import Home from "../pages/Home";
 
+// EXPERIMENT
+import Experiment from "../pages/secret/experiment";
+
 // SUPPORT
 import Help from "../pages/support/Help";
 import Settings from "../pages/support/Settings";
@@ -36,46 +39,48 @@ import InventoryList from "../pages/inventory/ViewInventory";
 import AddInventory from "../pages/inventory/AddInventory";
 
 const Routing = () => {
-	const location = useLocation();
+  const location = useLocation();
 
-	return (
-		<AnimatePresence mode="wait">
-				<Routes location={location} key={location.pathname}>
-					<Route path="/" element={<Home />} />
+  return (
+    <AnimatePresence mode="wait">
+      <Routes location={location} key={location.pathname}>
+        <Route path="/" element={<Home />} />
+        <Route path="/secret" element={<Experiment />} />
 
-					{/* SUPPORT */}
-					<Route path="/help" element={<Help />} />
-					<Route path="/settings" element={<Settings />} />
-					<Route path="/contact" element={<ContactUs />} />
+        {/* SUPPORT */}
+        <Route path="/help" element={<Help />} />
+        <Route path="/settings" element={<Settings />} />
+        <Route path="/contact" element={<ContactUs />} />
 
-					{/* PRODUCTS */}
-					<Route path="/products" element={<ProductList />} />
-					<Route path="/products/add" element={<AddProducts />} />
-					<Route path="/products/edit/:productId" element={<EditProducts />} />
+        {/* PRODUCTS */}
+        <Route path="/products" element={<ProductList />} />
+        <Route path="/products/add" element={<AddProducts />} />
+        <Route path="/products/edit/:productId" element={<EditProducts />} />
 
-					{/* VENDORS */}
-					<Route path="/vendors" element={<VendorList />} />
-					<Route path="/vendors/add" element={<AddVendors />} />
-					<Route path="/vendors/edit/:vendorId" element={<UpdateVendor />} />
+        {/* VENDORS */}
+        <Route path="/vendors" element={<VendorList />} />
+        <Route path="/vendors/add" element={<AddVendors />} />
+        <Route path="/vendors/edit/:vendorId" element={<UpdateVendor />} />
 
-					{/* TRANSPORT */}
-					<Route path="/transport" element={<TransportList />} />
-					<Route path="/transport/add" element={<AddTransport />} />
-					<Route path="/transport/edit/:transportId" element={<UpdateTransport />} />
+        {/* TRANSPORT */}
+        <Route path="/transport" element={<TransportList />} />
+        <Route path="/transport/add" element={<AddTransport />} />
+        <Route
+          path="/transport/edit/:transportId"
+          element={<UpdateTransport />}
+        />
 
-					{/* OFFERS */}
-					<Route path="/offers" element={<Offers />} />
-					<Route path="/offers/add" element={<AddOffer />} />
-					<Route path="/offers/edit/:offerId" element={<UpdateOffer />} />
-				
-					{/* INVENTORY */}
-					<Route path="/inventory" element={<InventoryList />} />
-					<Route path="/inventory/add" element={<AddInventory />} />
+        {/* OFFERS */}
+        <Route path="/offers" element={<Offers />} />
+        <Route path="/offers/add" element={<AddOffer />} />
+        <Route path="/offers/edit/:offerId" element={<UpdateOffer />} />
 
-				
-				</Routes>
-		</AnimatePresence>
-	);
+        {/* INVENTORY */}
+        <Route path="/inventory" element={<InventoryList />} />
+        <Route path="/inventory/add" element={<AddInventory />} />
+      </Routes>
+    </AnimatePresence>
+  );
 };
 
 export default Routing;
